Move score is now in 16 bits

+++ src/history.cpp
@@ -41,10 +41,10 @@ void HistoryTable::addNotCutoff(const Position& pos, const Move& move, const int
     butterfly[pos.getBoard(move.getFrom())][move.getTo()] += depth * depth;
 }
 
-int HistoryTable::getScore(const Position& pos, const Move& move) const
+int16_t HistoryTable::getScore(const Position& pos, const Move& move) const
 {
-    return ((history[pos.getBoard(move.getFrom())][move.getTo()] * 100) / 
-            (history[pos.getBoard(move.getFrom())][move.getTo()] + butterfly[pos.getBoard(move.getFrom())][move.getTo()] + 1));
+    return static_cast<int16_t>(((history[pos.getBoard(move.getFrom())][move.getTo()] * 100) / 
+            (history[pos.getBoard(move.getFrom())][move.getTo()] + butterfly[pos.getBoard(move.getFrom())][move.getTo()] + 1)));
 }
 
 void HistoryTable::age()
